# Reduce buffer allocations