… src/Microsoft.OData.Core/Microsoft.OData.Core.cs
… src/Microsoft.OData.Core/Microsoft.OData.Core.txt
… src/Microsoft.OData.Core/ODataConstants.cs
… src/Microsoft.OData.Core/ODataMessageWriterSettings.cs
… src/Microsoft.OData.Core/ODataMetadataOutputContext.cs
… src/Microsoft.OData.Core/ODataRawOutputContext.cs
… src/Microsoft.OData.Core/Parameterized.Microsoft.OData.Core.cs
… src/Microsoft.OData.Core/PublicAPI/net45/PublicAPI.Unshipped.txt
… src/Microsoft.OData.Core/PublicAPI/netcoreapp3.1/PublicAPI.Unshipped.txt
… src/Microsoft.OData.Core/PublicAPI/netstandard1.1/PublicAPI.Unshipped.txt
… src/Microsoft.OData.Core/PublicAPI/netstandard2.0/PublicAPI.Unshipped.txt
… test/FunctionalTests/Microsoft.OData.Core.Tests/Json/PooledByteBufferWriterTests.cs
… test/FunctionalTests/Microsoft.OData.Core.Tests/ODataMessageWriterSettingsTests.cs
… test/PerformanceTests/SerializationComparisonsTests/JsonWriterBenchmarks/Benchmarks.cs
… test/PerformanceTests/SerializationComparisonsTests/Lib/ODataMessageWriterAsyncPayloadWriter.cs
… test/PerformanceTests/SerializationComparisonsTests/Lib/ODataMessageWriterPayloadWriter.cs
… test/PublicApiTests/BaseLine/Microsoft.OData.PublicApi.net45.bsl
… test/PublicApiTests/BaseLine/Microsoft.OData.PublicApi.netstandard1.1.bsl
… test/PublicApiTests/BaseLine/Microsoft.OData.PublicApi.netstandard2.0.bsl