<!-- markdownlint-disable MD002 MD041 -->

### Issues

*This pull request fixes #2790 .*

### Description

This PR reduces allocations of buffer used by ODL's internal JSON writer:
- When using `JsonWriter`:
  - made the buffer size of `BufferedStream` configurable using `ODataMessageWriterSettings.BufferSize`
- When using the `ODataUtf8JsonWriter`:
  - we do not wrap the stream in a `BufferedStream` since `ODataUtf8JsonWriter` handles its own buffering
  - replaces the `ArrayBufferWriter` used by `ODataUtf8JsonWriter` with a custom `PooledByteBufferWriter` that rents buffers from the Array pool

The `PooledByteBufferWriter` was adapted from .NET's implementation: https://github.com/dotnet/runtime/blob/main/src/libraries/Common/src/System/Text/Json/PooledByteBufferWriter.cs

**Update**: Instead of hardcoding the default size to a lower value, I made it configurable by adding a `BufferSize` property to `ODataMessageWriterSettings`. However, this setting is only used when creating an asynchronous `JsonWriter`. It is not used when creating an instance of `ODataUtf8JsonWriter`. The reason is because `IStreamBasedJsonWriterFactory.CreateJsonWriter` unfortunately does not expose a parameter for passing the buffer size.

I also made changes to the `SerializationComparisonTests` benchmarks to ensure that the writers are disposed after the payload has been written. This required refactoring the `JsonWriterBenchmarks` such that a new stream is used for each write call. The benchmarks can't re-use the stream cause the stream gets disposed. The changes make the benchmarks arguably cleaner even though each benchmark call is doing a bit more work than just writing the payload. This means that a comparison of benchmark results between this PR and the master branch is no longer an apples to apples comparison.

 When using `JsonWriter`, I had initially reduced the `BufferedStream` buffer size from 84k to 16k but I observed mixed results from the benchmarks, no strong evidence of improvement, and in some cases there was a regression. As a result I retained the original 84k buffer size but made it configurable via `ODataMessageWriterSettings.BufferSize`

Generally, I didn't see a lot of conclusive improvement in latency. But it doesn't seem to make latency worse. If it improves memory, without degrading latency, I think that's a good enough improvement to have.

### Benchmarks

I used the `TestServer` project in `SerializationComparisonsTests` to compare the heap allocations and latency/throughput before and after the changes.

#### Memory allocations

I set the count to 2000 to get the payload size above 400KB, which is a relatively large payload size, but matches a sample payload size of one of the customers.

Command used for testing:
-  Warm up request `bombardier -l -c 1 -n 1 https://localhost:7120/customers/ODataMessageWriter-Async`
- Concurrent requests: `bombardier -l -c 10 -n 50 https://localhost:7120/customers/ODataMessageWriter-Async`

**`JsonWriter` Before**

![image](https://github.com/OData/odata.net/assets/8460169/c3fffb6e-323a-4aab-9a9c-a8fbd414a9ca)

![image](https://github.com/OData/odata.net/assets/8460169/18fec3a4-ab68-4f38-9de3-c109a4820aa4)

**`JsonWriter` After**

![image](https://github.com/OData/odata.net/assets/8460169/b9694850-52ce-49b8-9a12-a432f346501a)

![image](https://github.com/OData/odata.net/assets/8460169/ae77f440-ebc1-4756-b87f-76e94c7bf824)


**`Utf8JsonWriter` Before**

![image](https://github.com/OData/odata.net/assets/8460169/f495b825-9b4e-450e-aa8f-00b6ee68ecc7)

![image](https://github.com/OData/odata.net/assets/8460169/ea035463-2c3a-49f5-bdbb-e1fedf696cb9)


**`Utf8JsonWriter` After**

![image](https://github.com/OData/odata.net/assets/8460169/d4c4bd68-9ad3-49ec-be23-e9c9433da236)

![image](https://github.com/OData/odata.net/assets/8460169/5f5d3f8b-c336-4772-8e9f-82ed48bb420d)


#### Request metrics

Command used for testing:

- Sequential request latency: `bombardier -l -c 1 -n 100 https://localhost:7120/customers/ODataMessageWriter-Async?count=2000`
- Concurrent requests: `bombardier -l -c 50 -d 1m https://localhost:7120/customers/ODataMessageWriter-Async?count=2000`
- High load (timeouts expected): `Utf8JsonWriter`: `bombardier -l -c 100 -d 1m https://localhost:7120/customers/ODataMessageWriter-Async?

**`JsonWriter` before**

- Sequential requests:
Statistics        Avg      Stdev        Max
  Reqs/sec        12.76      18.91      82.50
  Latency       76.91ms    45.08ms   406.04ms
  Latency Distribution
     50%    64.69ms
     75%    66.31ms
     90%    75.04ms
     95%   174.68ms
     99%   220.88ms
  HTTP codes:
    1xx - 0, 2xx - 100, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     7.43MB/s

- 50 connections concurrent requests:

Statistics        Avg      Stdev        Max
  Reqs/sec        63.07     145.61    3849.11
  Latency         0.94s    98.78ms      1.67s
  Latency Distribution
     50%      0.94s
     75%      1.00s
     90%      1.06s
     95%      1.11s
     99%      1.33s
  HTTP codes:
    1xx - 0, 2xx - 3193, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    30.32MB/s

- 100 connections

Statistics        Avg      Stdev        Max
  Reqs/sec        49.50     117.08    1467.94
  Latency         2.28s      1.88s     13.83s
  Latency Distribution
     50%      1.75s
     75%      1.87s
     90%      4.13s
     95%      6.53s
     99%     12.67s
  HTTP codes:
    1xx - 0, 2xx - 2639, 3xx - 0, 4xx - 0, 5xx - 0
    others - 31

**`JsonWriter` after**

- Sequential requests:

Statistics        Avg      Stdev        Max
  Reqs/sec        12.68      19.03      88.51
  Latency       78.80ms    40.66ms   426.61ms
  Latency Distribution
     50%    68.24ms
     75%    71.52ms
     90%    81.92ms
     95%   140.85ms
     99%   160.52ms
  HTTP codes:
    1xx - 0, 2xx - 100, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     7.29MB/s

- 50 connections, concurrent requests:

Statistics        Avg      Stdev        Max
  Reqs/sec        66.39     139.43    3841.72
  Latency         0.90s    84.35ms      1.28s
  Latency Distribution
     50%      0.90s
     75%      0.95s
     90%      1.00s
     95%      1.04s
     99%      1.13s
  HTTP codes:
    1xx - 0, 2xx - 3361, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    31.99MB/s

- 100 connections

Statistics        Avg      Stdev        Max
  Reqs/sec        53.45     179.59    5419.08
  Latency         2.36s      2.02s     13.55s
  Latency Distribution
     50%      1.75s
     75%      1.89s
     90%      4.22s
     95%      6.94s
     99%     12.67s
  HTTP codes:
    1xx - 0, 2xx - 2546, 3xx - 0, 4xx - 0, 5xx - 0
    others - 34

**`Utf8JsonWriter` before**
- Sequential requests

Statistics        Avg      Stdev        Max
  Reqs/sec        15.01      19.06      63.33
  Latency       66.21ms    34.34ms   345.19ms
  Latency Distribution
     50%    56.89ms
     75%    59.79ms
     90%    70.13ms
     95%   109.54ms
     99%   159.28ms
  HTTP codes:
    1xx - 0, 2xx - 100, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     8.89MB/s

- 50 connections

Statistics        Avg      Stdev        Max
  Reqs/sec        76.40     160.26    3675.45
  Latency      818.42ms    96.34ms      1.23s
  Latency Distribution
     50%   807.87ms
     75%      0.88s
     90%      0.95s
     95%      1.01s
     99%      1.10s
  HTTP codes:
    1xx - 0, 2xx - 3681, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    35.96MB/s

- 100 connections

Statistics        Avg      Stdev        Max
  Reqs/sec        71.17     131.48    2669.99
  Latency         1.65s   391.15ms     11.73s
  Latency Distribution
     50%      1.61s
     75%      1.70s
     90%      1.82s
     95%      1.89s
     99%      4.12s
  HTTP codes:
    1xx - 0, 2xx - 3663, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    36.03MB/s

**`Utf8JsonWriter` after**

- Sequential requests

Statistics        Avg      Stdev        Max
  Reqs/sec        17.12      62.12    1135.20
  Latency       70.11ms    36.60ms   347.11ms
  Latency Distribution
     50%    57.99ms
     75%    61.92ms
     90%   107.97ms
     95%   138.66ms
     99%   156.07ms
  HTTP codes:
    1xx - 0, 2xx - 100, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     8.40MB/s

- 50 connections

Statistics        Avg      Stdev        Max
  Reqs/sec        75.81     147.23    3844.68
  Latency      792.53ms    80.07ms      1.33s
  Latency Distribution
     50%   790.36ms
     75%   845.36ms
     90%      0.90s
     95%      0.93s
     99%      1.00s
  HTTP codes:
    1xx - 0, 2xx - 3802, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    37.13MB/s

-100 connections

Statistics        Avg      Stdev        Max
  Reqs/sec        73.47     209.24    5402.75
  Latency         1.71s   544.48ms      9.21s
  Latency Distribution
     50%      1.63s
     75%      1.74s
     90%      1.86s
     95%      1.97s
     99%      4.21s
  HTTP codes:
    1xx - 0, 2xx - 3533, 3xx - 0, 4xx - 0, 5xx - 0
    others - 1
  Errors:
    tls handshake timed out - 1
  Throughput:    35.44MB/s

#### Writer benchmarks

**Before**
|           Method |                                           WriterName |       Mean |      Error |     StdDev |     Median |      Gen 0 |     Gen 1 |  Allocated |
|----------------- |----------------------------------------------------- |-----------:|-----------:|-----------:|-----------:|-----------:|----------:|-----------:|
| WriteToFileAsync |                                   ODataMessageWriter | 380.980 ms |  7.4531 ms |  6.6070 ms | 379.283 ms | 53000.0000 |         - | 224,003 KB |
| WriteToFileAsync |                             ODataMessageWriter-Async | 684.025 ms |  9.7998 ms |  9.1667 ms | 681.923 ms | 63000.0000 | 1000.0000 | 266,767 KB |
| WriteToFileAsync |                    ODataMessageWriter-Utf8JsonWriter | 353.918 ms |  6.9733 ms |  5.8231 ms | 354.351 ms | 51000.0000 | 1000.0000 | 217,429 KB |
| WriteToFileAsync |              ODataMessageWriter-Utf8JsonWriter-Async | 620.257 ms | 12.2123 ms | 15.8794 ms | 613.840 ms | 51000.0000 | 1000.0000 | 218,772 KB |

**After**

|           Method |                                           WriterName |       Mean |      Error |     StdDev |     Median |      Gen 0 |     Gen 1 |  Allocated |
|----------------- |----------------------------------------------------- |-----------:|-----------:|-----------:|-----------:|-----------:|----------:|-----------:|
| WriteToFileAsync |                                   ODataMessageWriter | 399.640 ms |  7.9563 ms |  9.4714 ms | 396.437 ms | 53000.0000 |         - | 224,018 KB |
| WriteToFileAsync |                             ODataMessageWriter-Async | 704.245 ms | 13.5573 ms | 14.5062 ms | 698.330 ms | 63000.0000 | 1000.0000 | 266,773 KB |
| WriteToFileAsync |                    ODataMessageWriter-Utf8JsonWriter | 378.397 ms |  3.2468 ms |  2.8782 ms | 378.249 ms | 51000.0000 |         - | 217,344 KB |
| WriteToFileAsync |              ODataMessageWriter-Utf8JsonWriter-Async | 627.557 ms | 12.4976 ms | 20.8807 ms | 623.245 ms | 51000.0000 |         - | 218,138 KB |

### Checklist (Uncheck if it is not completed)

- [x] *Test cases added*
- [x] *Build and test with one-click build and test script passed*