@habbes 2023-11-15
Based on the results of the benchmarks, I'm reconsidering whether it's a good idea to reduce the buffer size of `BufferedStream` for the default `JsonWriter`.